+++ packages/react-impression/src/components/Select/Select.js
@@ -36,6 +36,7 @@ export default class Select extends React.PureComponent {
       currentPlaceholder: this.props.placeholder,
       queryText: '', // 搜索字段
       showClear: false,
+      scrollDistance: 10, // 距离底部多远触发onScrollBottom，默认值为10
     }
 
     this.state = {
@@ -127,6 +128,20 @@ export default class Select extends React.PureComponent {
      * 弹出层宽度伸缩方式
      */
     stretch: PropTypes.oneOf(['sameWidth', 'auto']),
+
+    /**
+     * 指定selectOption渲染的dom节点
+     */
+    container: PropTypes.oneOfType([PropTypes.node, PropTypes.undefined]),
+    /**
+     * 滚动到底部
+     *
+     */
+    onScrollBottom: PropTypes.func,
+    /**
+     * 下拉加载loading文案
+     */
+    loadingText: PropTypes.string,
   }
   static defaultProps = {
     disabled: false,
@@ -137,6 +152,38 @@ export default class Select extends React.PureComponent {
 
   componentDidMount() {
     this.handleInit()
+    this.handleListnerScroll()
+  }
+
+  /**
+   * 处理下拉加载
+   */
+  handleListnerScroll = () => {
+    if (
+      this.props.onScrollBottom &&
+      typeof this.props.onScrollBottom === 'function'
+    ) {
+      this.addScrollListener()
+    }
+  }
+
+  handleScroll = e => {
+    if (
+      e.target.scrollHeight - e.target.scrollTop - e.target.offsetHeight <=
+      this.state.scrollDistance
+    ) {
+      this.props.onScrollBottom()
+    }
+  }
+
+  addScrollListener = () => {
+    this.selectInner &&
+      this.selectInner.addEventListener('scroll', this.handleScroll)
+  }
+
+  removeScrollListener = () => {
+    this.selectInner &&
+      this.selectInner.removeEventListener('scroll', this.handleScroll)
   }
 
   /**
@@ -182,7 +229,9 @@ export default class Select extends React.PureComponent {
     if (!multiple) {
       let selectedItem = {}
       // 非远程搜索
-      if (optionList.length > 0) {
+      if (remoteMethod && originValue && originValue.constructor === Object) {
+        selectedItem = originValue
+      } else if (optionList.length > 0) {
         selectedItem =
           optionList.find(option => {
             return option.value === originValue
@@ -202,10 +251,9 @@ export default class Select extends React.PureComponent {
     if (multiple) {
       let selectList = []
       const [originalItem = ''] = originValue || []
-      if (
-        !remoteMethod ||
-        (remoteMethod && originalItem && originalItem.constructor !== Object)
-      ) {
+      if (originalItem && originalItem.constructor === Object) {
+        selectList = originValue
+      } else {
         originValue &&
           originValue.length > 0 &&
           optionList.length > 0 &&
@@ -215,9 +263,6 @@ export default class Select extends React.PureComponent {
             })
             item && selectList.push(item)
           })
-      }
-      if (originalItem && originalItem.constructor === Object) {
-        selectList = originValue
       }
       dataToSet = {
         selectedItem: selectList || [],
@@ -378,7 +423,13 @@ export default class Select extends React.PureComponent {
    * @param result
    */
   selectOptionHandle(result) {
-    const { onChange, value, multiple } = this.props
+    const {
+      onChange,
+      value,
+      multiple,
+      remoteMethod,
+      onScrollBottom,
+    } = this.props
     const { options, selectedOptions, selectedItem } = this.state
     const originValue = this.isPuppet ? value : this.state.value
     if (multiple) {
@@ -405,10 +456,22 @@ export default class Select extends React.PureComponent {
         }
       )
     } else {
+      if (
+        !multiple &&
+        this.state.queryText &&
+        !!remoteMethod &&
+        !!onScrollBottom
+      ) {
+        remoteMethod('')
+      }
       this.setState(
         {
           selectText: multiple ? '' : result.name,
-          queryText: multiple ? '' : result.name,
+          queryText: multiple
+            ? !!remoteMethod && !!onScrollBottom
+              ? this.state.queryText
+              : ''
+            : result.name,
         },
         () => {
           onChange &&
@@ -428,6 +491,12 @@ export default class Select extends React.PureComponent {
    */
   componentWillUnmount() {
     window.cancelAnimationFrame(this.requestId)
+    if (
+      this.props.onScrollBottom &&
+      typeof this.props.onScrollBottom === 'function'
+    ) {
+      this.removeScrollListener()
+    }
   }
 
   UNSAFE_componentWillReceiveProps(nextProps) {
@@ -588,6 +657,10 @@ export default class Select extends React.PureComponent {
   hideOptionsHandler = popupVisible => {
     if (!popupVisible && this.state.showOption) {
       const { selectText } = this.state
+      const { remoteMethod, onScrollBottom } = this.props
+      if (this.state.queryText && !!remoteMethod && !!onScrollBottom) {
+        remoteMethod('')
+      }
       this.setState({ showOption: false, queryText: selectText }, () => {
         this.selectInner.scrollTop = 0
       })
@@ -606,6 +679,8 @@ export default class Select extends React.PureComponent {
       clearable,
       size,
       stretch,
+      container,
+      outsideDisabled = false,
     } = this.props
     const {
       showOption,
@@ -652,9 +727,14 @@ export default class Select extends React.PureComponent {
                 {this.emptyText && (
                   <p className='select-empty'>{this.emptyText}</p>
                 )}
+                {this.props.loadingText && (
+                  <p className='select-empty'>{this.props.loadingText}</p>
+                )}
               </ul>
             </div>
           }
+          container={container}
+          outsideDisabled={outsideDisabled}
         >
           <div
             style={style}
@@ -706,8 +786,8 @@ export default class Select extends React.PureComponent {
               <input
                 type='text'
                 value={queryText}
-                readOnly={!searchable || (searchable && !showOption)}
-                placeholder={currentPlaceholder || '请选择'}
+                readOnly={!searchable}
+                placeholder={currentPlaceholder || placeholder}
                 disabled={disabled}
                 className='select-selection'
                 onChange={e => {